+++ apps/inp-testing/src/routes/ui3.tsx
@@ -1,8 +1,10 @@
 import {useState} from 'react'
 import {
+  Box,
   Button,
   Card,
   Checkbox,
+  Dialog,
   Grid,
   Heading,
   Inline,
@@ -73,6 +75,27 @@ function SingleControlSection() {
         </Inline>
       </Stack>
     </Card>
+  )
+}
+
+function DialogSection() {
+  const [dialogOpen, setDialogOpen] = useState(false)
+  return (
+    <Box marginY={5}>
+      <Button text="Open Dialog" onClick={() => setDialogOpen(true)} />
+      {dialogOpen && (
+        <Dialog
+          animate={true}
+          id="inpDialog"
+          onClose={() => setDialogOpen(false)}
+          header="Dialog header"
+        >
+          <Box padding={4}>
+            <Text>The text of the Modal</Text>
+          </Box>
+        </Dialog>
+      )}
+    </Box>
   )
 }
 
@@ -272,6 +295,7 @@ function Ui3() {
       <Stack gap={4}>
         <Heading>UI 3</Heading>
         <SingleControlSection />
+        <DialogSection />
         <SelectAllSection />
         <ToneToggleSection />
         <PanelSwapSection />